requirements.txt
streamlit
python-docx
nltk
PyPDF2
